[utils] Move ``eos::qnp`` classes into a separate header file
